# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vehicle Rental API",
        "description": "API для управления арендой транспортных средств",
        "contact": {
            "email": "vahagn99ghukasyan@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/vehicles": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Список всех ТС",
                "operationId": "bcdde16293935e06d4c75c4339470415",
                "responses": {
                    "200": {
                        "description": "Список ТС",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Vehicle"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/vehicles/{id}": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Получить ТС по ID",
                "operationId": "17b29272b3e37d1a2c440facdddecc7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ТС найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehicle"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/vehicles/{id}/status": {
            "patch": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Обновить статус ТС",
                "operationId": "7f30435a4b87830cc045be1b9a542311",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Статус обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "updated": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/vehicles/{id}/location": {
            "patch": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Обновить координаты ТС",
                "operationId": "a656587485077fbf886b199ed22c5217",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 40.7128
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -74.006
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Локация обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehicle"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Vehicle": {
                "title": "Vehicle",
                "description": "Автомобиль",
                "properties": {
                    "id": {
                        "description": "Идентификатор",
                        "type": "integer",
                        "format": "int64"
                    },
                    "model_id": {
                        "description": "Идентификатор модели",
                        "type": "integer",
                        "format": "int64"
                    },
                    "vin": {
                        "description": "VIN номер",
                        "type": "string"
                    },
                    "number_plate": {
                        "description": "Номерной знак",
                        "type": "string"
                    },
                    "status": {
                        "description": "Статус автомобиля",
                        "type": "string"
                    },
                    "location": {
                        "description": "Местоположение автомобиля",
                        "type": "object",
                        "example": {
                            "latitude": 40.7128,
                            "longitude": -74.006
                        }
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Овторизация с помощью токена",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}